Fix UiDateTime build and demo toggle composition
UiDateTime.h included <CtrlLib/DateTimeCtrl.h> at global scope, but that
header is unguarded and only valid inside namespace Upp. Pull the U++
core headers in first so the control compiles under BLITZ.

Also:
- Cast SUNDAY/SATURDAY to int in SetFirstDayOfWeek so minmax sees one type.
- Make the event test date-independent: the default value is today, so
  committing today's date fired no change/action. Set an explicit earlier
  value before registering counters.
- Demo: replace UiCompositeToggle usage (targeted for removal) with plain
  UiToggle plus a separate UiLabel in front of each switch.

UiDateTimeRunTests: Checks 25, Fails 0 (run twice). Demo builds, links
and launches without warnings.

+++ Ui/UiDateTime.cpp
@@ -427,7 +427,7 @@ UiDateTime& UiDateTime::ClearRange()
 
 UiDateTime& UiDateTime::SetFirstDayOfWeek(int day)
 {
-    first_day_ = minmax(day, SUNDAY, SATURDAY);
+    first_day_ = minmax(day, (int)SUNDAY, (int)SATURDAY);
     date_popup_.FirstDay(first_day_);
     datetime_popup_.calendar.FirstDay(first_day_);
     return *this;
